Set SELinux to permissive for operation; set back
Some of the STIG changes require SELinux to be off for the changes to be
made by the autobot user.

This may be false and we may need to back this out later, but this is
how I got it working for now.

It sets SELinux back to enforcing when it's done.